Package miou

Sub-indexes

backoff

sync

Library miou

This library exposes the following toplevel modules:

Miou

Miou_fmt

Miou_gen

Miou_logs

Miou_pqueue This module implements a priority queue based on a minimal binary heap. The heap is implemented as a dynamic array, taken from the module vector.

Miou_queue

Miou_sequence Mutable sequence of elements.

Miou_state

Miou_vector Vectors (aka resizable arrays, growing arrays, dynamic arrays, etc.)

Library miou.backoff

The entry point of this library is the module: Miou_backoff.

Library miou.sync

The entry point of this library is the module: Miou_sync.